Moved some package in their appropriate place.
httpclient is now under arduino/httpclient.
Some configuration values have been moved under 'configuration' package,
and the related subroutines have been refactored.
