+++ arduino/httpclient/httpclient_test.go
