--- httpclient/httpclient_transport.go
@@ -1,41 +1,0 @@
-// This file is part of arduino-cli.
-//
-// Copyright 2020 ARDUINO SA (http://www.arduino.cc/)
-//
-// This software is released under the GNU General Public License version 3,
-// which covers the main part of arduino-cli.
-// The terms of this license can be found at:
-// https://www.gnu.org/licenses/gpl-3.0.en.html
-//
-// You can be released from the requirements of the above licenses by purchasing
-// a commercial license. Buying such a license is mandatory if you want to
-// modify or otherwise use the software for commercial activities involving the
-// Arduino software without disclosing the source code of your own applications.
-// To purchase a commercial license, send an email to license@arduino.cc.
-
-package httpclient
-
-import "net/http"
-
-type httpClientRoundTripper struct {
-	transport http.RoundTripper
-	config    *Config
-}
-
-func newHTTPClientTransport(config *Config) http.RoundTripper {
-	proxy := http.ProxyURL(config.Proxy)
-
-	transport := &http.Transport{
-		Proxy: proxy,
-	}
-
-	return &httpClientRoundTripper{
-		transport: transport,
-		config:    config,
-	}
-}
-
-func (h *httpClientRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.Header.Add("User-Agent", h.config.UserAgent)
-	return h.transport.RoundTrip(req)
-}
